refactor: create model files #4
