+++ server/data/database/constants.js
@@ -1,0 +1,4 @@
+//
+// Constants
+//
+export const VIEWER_ID = 'me';
